Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main() {
-    int a , b, result;
+    double a , b, result;
     char op;
-    scanf("%d %d %c", &a, &b, &op);
+    scanf("%lf %lf %c", &a, &b, &op);
     switch(op){
         case '+': result = a+b;
         case '-': reuslt = a-b;
